fix and improve fm commands is_visible method
The logic use to be wrong with the menu_without_distraction (which
honestly, no one should set to false anyway). It also now prints stuff
to the console when it detects an internal error. I don't think this is
the best way to report the error, but i don't want to nag the user
with stupid pop ups, so i try to guess the best behavior when that
happens.