Warning for user of the auto_close_empty_groups
It's better than setting an "install update" in messages, because I
never bother to read those (so I'm guessing people don't) (idea: the
problem is that it's hard to emphasize stuff. There should be a syntax
definition to highlight important changes), and because it only affect
people who *we know* it is going to affect. It's not a fat disgusting
promotion to uninterested programmers who just want to get stuff done.

So, the setting is officially removed from FileManager.

I'm happy with it :-)

Also, I fixed some indentation is the only interesting part of the
install message. I probably should fix it up too.