Improve submodule reloading
This commit provides a change to clear the sys.modules cache before
importing any local package/module. This causes python to reload all
modules as soon as the FileManager.py is reloaded/imported by ST.

It is to ensure fluent plugin updates even if global variables or
module names change.

Nothing else is needed to instantly update all submodules during
development but saving the FileManager.py

This strategy is used by `A File Icon` and `GitGutter` for instance.