Apply requested review changes
OK, a `num_paths` variable is considered useless even though it avoids
duplicated evaluations of `self.path`, ... ? Lets remove it and the
other useless temporary variables, which are used once only.

The first 2 quick panel items can easily be created in-place.
The number of "format()" calls is reduced for cheaper item creation.

There is no need to display the number of items to delete in both items. So we can use the description line of "Cancel All" command to
describe how to remove individual items from the list of "to-delete"
files.