Fix FmDelete command (#64)
* Fix FmDelete command

Commit 5f1b64c8d7eee6163b78d4b4f014b7a461a5e7a9 intends to enable us to
individually select files to delete from the quick panel.

1. The following changes are made to the new functionality:

   a) Don't close a related view, as the file won't be deleted, if
      removed from the quick panel.
   b) Don't refresh the sidebar folder as nothing is changed, if a
      file is removed from the quick panel.
      (Also avoids refreshing the sidbear twice)

2. Ensures the public `run()` to be the first method in the class.
3. Avoids calling `len(self.path)` multiple times.
4. Use list arithmetic to create the quick panel items.

* Apply requested review changes

OK, a `num_paths` variable is considered useless even though it avoids
duplicated evaluations of `self.path`, ... ? Lets remove it and the
other useless temporary variables, which are used once only.

The first 2 quick panel items can easily be created in-place.
The number of "format()" calls is reduced for cheaper item creation.

There is no need to display the number of items to delete in both items. So we can use the description line of "Cancel All" command to
describe how to remove individual items from the list of "to-delete"
files.